+++ HISTORY.md
@@ -53,6 +53,7 @@
 
 ### 2022-05-29
 * blocks, threads: new "slots" choice in the OF BLOCK block-attribute reporter's dropdown
+* blocks, threads: programmatically change slot shapes in custom blocks
 
 ### 2022-05-28
 * new extension primitive for importing a costume from a url